Add files via upload
This script requests user input and solves for the number of Collatz sequences needed to evaluate to 1 from numbers 1 through the integer input provided. It will then provide as an output the maximum number of sequences in the selected range needed to evaluate to 1, as well as the number(s) at which this takes place.